src/docvector/vector_stores.py
"""
Vector store implementations for document storage and retrieval.
"""

from abc import ABC, abstractmethod
from typing import List, Optional, Dict, Any, Tuple
from .types import Document, Chunk

class BaseVectorStore(ABC):
    """Base class for vector stores."""
    
    @abstractmethod
    def add_document(self, document: Document) -> str:
        """Add a document to the vector store."""
        pass
    
    @abstractmethod
    def add_documents(self, documents: List[Document]) -> List[str]:
        """Add multiple documents to the vector store."""
        pass
    
    @abstractmethod
    def search(self, query: str, limit: int = 5) -> List[Document]:
        """Search for similar documents."""
        pass
    
    @abstractmethod
    def delete_document(self, document_id: str) -> bool:
        """Delete a document from the vector store."""
        pass
    
    @abstractmethod
    def clear(self) -> None:
        """Clear all documents from the vector store."""
        pass

class QdrantStore(BaseVectorStore):
    """Qdrant vector store implementation."""
    
    def __init__(self, url: str, collection_name: str = "documents"):
        self.url = url
        self.collection_name = collection_name
        # TODO: Initialize Qdrant client
    
    def add_document(self, document: Document) -> str:
        """Add a document to Qdrant."""
        # TODO: Implement Qdrant document addition
        return "dummy_id"
    
    def add_documents(self, documents: List[Document]) -> List[str]:
        """Add multiple documents to Qdrant."""
        return [self.add_document(doc) for doc in documents]
    
    def search(self, query: str, limit: int = 5) -> List[Document]:
        """Search for similar documents in Qdrant."""
        # TODO: Implement Qdrant search
        return []
    
    def delete_document(self, document_id: str) -> bool:
        """Delete a document from Qdrant."""
        # TODO: Implement Qdrant document deletion
        return True
    
    def clear(self) -> None:
        """Clear all documents from Qdrant."""
        # TODO: Implement Qdrant collection clearing
        pass

class WeaviateStore(BaseVectorStore):
    """Weaviate vector store implementation."""
    
    def __init__(self, url: str, class_name: str = "Document"):
        self.url = url
        self.class_name = class_name
        # TODO: Initialize Weaviate client
    
    def add_document(self, document: Document) -> str:
        """Add a document to Weaviate."""
        # TODO: Implement Weaviate document addition
        return "dummy_id"
    
    def add_documents(self, documents: List[Document]) -> List[str]:
        """Add multiple documents to Weaviate."""
        return [self.add_document(doc) for doc in documents]
    
    def search(self, query: str, limit: int = 5) -> List[Document]:
        """Search for similar documents in Weaviate."""
        # TODO: Implement Weaviate search
        return []
    
    def delete_document(self, document_id: str) -> bool:
        """Delete a document from Weaviate."""
        # TODO: Implement Weaviate document deletion
        return True
    
    def clear(self) -> None:
        """Clear all documents from Weaviate."""
        # TODO: Implement Weaviate class clearing
        pass

class MilvusStore(BaseVectorStore):
    """Milvus vector store implementation."""
    
    def __init__(self, host: str, port: int, collection_name: str = "documents"):
        self.host = host
        self.port = port
        self.collection_name = collection_name
        # TODO: Initialize Milvus client
    
    def add_document(self, document: Document) -> str:
        """Add a document to Milvus."""
        # TODO: Implement Milvus document addition
        return "dummy_id"
    
    def add_documents(self, documents: List[Document]) -> List[str]:
        """Add multiple documents to Milvus."""
        return [self.add_document(doc) for doc in documents]
    
    def search(self, query: str, limit: int = 5) -> List[Document]:
        """Search for similar documents in Milvus."""
        # TODO: Implement Milvus search
        return []
    
    def delete_document(self, document_id: str) -> bool:
        """Delete a document from Milvus."""
        # TODO: Implement Milvus document deletion
        return True
    
    def clear(self) -> None:
        """Clear all documents from Milvus."""
        # TODO: Implement Milvus collection clearing
        pass

class PineconeStore(BaseVectorStore):
    """Pinecone vector store implementation."""
    
    def __init__(self, api_key: str, environment: str, index_name: str):
        """
        Initialize Pinecone vector store.
        
        Args:
            api_key (str): Pinecone API key
            environment (str): Pinecone environment (e.g., 'us-west1-gcp')
            index_name (str): Name of the Pinecone index
        """
        try:
            import pinecone
            pinecone.init(api_key=api_key, environment=environment)
            self.index = pinecone.Index(index_name)
        except ImportError:
            raise ImportError("Please install pinecone-client: pip install pinecone-client")
    
    def add_document(self, document: Document) -> None:
        """Add a document to Pinecone."""
        vectors = []
        for i, chunk in enumerate(document.chunks):
            if chunk.embedding is not None:
                vector_id = f"{document.metadata.source}_{i}"
                metadata = {
                    "title": document.metadata.title,
                    "source": document.metadata.source,
                    "content_type": document.metadata.content_type,
                    "chunk_index": i,
                    "text": chunk.text
                }
                vectors.append((vector_id, chunk.embedding, metadata))
        
        if vectors:
            self.index.upsert(vectors=vectors)
    
    def search(self, query: str, top_k: int = 3) -> List[Tuple[Document, float]]:
        """Search for similar documents in Pinecone."""
        # Note: In a real implementation, you would need to embed the query first
        # This is a placeholder implementation
        results = []
        return results
    
    def delete_document(self, document_id: str) -> None:
        """Delete a document from Pinecone."""
        # Delete all vectors with the document prefix
        self.index.delete(filter={"source": document_id})
    
    def clear(self) -> None:
        """Clear all documents from Pinecone."""
        self.index.delete(delete_all=True) 