pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docvector"
version = "0.1.0"
description = "DocVector: Intelligent Document Processing & Knowledge Management System"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "python-dotenv>=1.0.0",
    "mistralai>=0.0.7",
    "qdrant-client>=1.6.0",
    "weaviate-client>=3.25.2",
    "pymilvus>=2.3.0",
    "pinecone-client>=2.2.4",
    "python-magic>=0.4.27",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "beautifulsoup4>=4.12.0",
    "markdown>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/jabbir-doodle/DocVector-"
Documentation = "https://github.com/jabbir-doodle/DocVector-/blob/main/README.md"
Repository = "https://github.com/jabbir-doodle/DocVector-.git"
Issues = "https://github.com/jabbir-doodle/DocVector-/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/docvector"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src/docvector"

[tool.coverage.run]
source = ["src/docvector"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "raise ImportError",
]

[tool.hatch.metadata]
allow-direct-references = true 